rename cli.same -> sameproject so that poetry publish works
