+++ backends/kubeflow/render.py
@@ -1,7 +1,7 @@
 from cli import same
 from objects.step import Step
 from pathlib import Path
-from cli.same import helpers
+from sameproject import helpers
 from jinja2 import Environment, FileSystemLoader, select_autoescape
 
 from kubernetes import client, config
