+++ clients/context.py
@@ -3,5 +3,5 @@ sys.path.insert(0, "..")
 
 
 from objects.step import Step
-from cli.same.program.compile import notebook_processing
+from sameproject.program.compile import notebook_processing
 from backends.durable_functions.constants import EXECUTE_WORKFLOW_ACTIVITY_NAME
