+++ sameproject/stdlib.py
